+++ packages/development-edition-engine-react/gulpfile.js
@@ -33,33 +33,33 @@ function serve() {
     });
 }
 
-gulp.task('patternlab:version', function() {
+gulp.task('patternlab:version', function () {
   patternlab.version();
 });
 
-gulp.task('patternlab:help', function() {
+gulp.task('patternlab:help', function () {
   patternlab.help();
 });
 
-gulp.task('patternlab:patternsonly', function() {
+gulp.task('patternlab:patternsonly', function () {
   patternlab.patternsonly(config.cleanPublic);
 });
 
-gulp.task('patternlab:liststarterkits', function() {
+gulp.task('patternlab:liststarterkits', function () {
   patternlab.liststarterkits();
 });
 
-gulp.task('patternlab:loadstarterkit', function() {
+gulp.task('patternlab:loadstarterkit', function () {
   patternlab.loadstarterkit(argv.kit, argv.clean);
 });
 
-gulp.task('patternlab:build', function() {
+gulp.task('patternlab:build', function () {
   build().then(() => {
     // do something else when this promise resolves
   });
 });
 
-gulp.task('patternlab:serve', function() {
+gulp.task('patternlab:serve', function () {
   serve().then(() => {
     // do something else when this promise resolves
   });