+++ packages/live-server/test/htpasswd.js
@@ -7,22 +7,17 @@ var liveServer = require('..').start({
   htpasswd: path.join(__dirname, 'data', 'htpasswd-test'),
 });
 
-describe('htpasswd tests', function() {
-  it('should respond with 401 since no password is given', function(done) {
-    request(liveServer)
-      .get('/')
-      .expect(401, done);
+describe('htpasswd tests', function () {
+  it('should respond with 401 since no password is given', function (done) {
+    request(liveServer).get('/').expect(401, done);
   });
-  it('should respond with 401 since wrong password is given', function(done) {
+  it('should respond with 401 since wrong password is given', function (done) {
     request(liveServer)
       .get('/')
       .auth('test', 'not-real-password')
       .expect(401, done);
   });
-  it('should respond with 200 since correct password is given', function(done) {
-    request(liveServer)
-      .get('/')
-      .auth('test', 'test')
-      .expect(200, done);
+  it('should respond with 200 since correct password is given', function (done) {
+    request(liveServer).get('/').auth('test', 'test').expect(200, done);
   });
 });