+++ packages/uikit-workshop/src/scripts/lit-components/pl-toggle-theme/pl-toggle-theme.js
@@ -72,7 +72,7 @@ class ThemeToggle extends LitElement {
       <pl-button
         class="pl-c-tools__action pl-c-toggle-theme__action"
         title="Switch Theme"
-        @click="${_ => store.dispatch(updateThemeMode(toggleThemeMode))}"
+        @click="${(_) => store.dispatch(updateThemeMode(toggleThemeMode))}"
       >
         Switch Theme